fix(profile): preserve avg correct on ungraded show days
Live season-stats fallback keeps careerCorrectSlots, and the finalized-show
heuristic excludes today so calendar show days do not false-stale rollups.
Also move Your stats definitions into Standings-style Info tooltips.

Closes #635 Slice 0; watermark Slices 1–2 remain on the issue.
