+++ src/features/profile/api/profileSeasonStats.test.js
@@ -1,6 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
-import { readMaterializedSeasonStats } from './profileSeasonStats';
+import {
+  deriveLatestFinalizedShow,
+  readMaterializedSeasonStats,
+  withCareerCorrectSlots,
+} from './profileSeasonStats';
 
 /**
  * #244 short-circuit decision logic. Kept as a pure predicate so the
@@ -100,5 +104,87 @@ describe('readMaterializedSeasonStats', () => {
         '2026-04-23'
       )
     ).toEqual({ totalPoints: 0, shows: 0, wins: 0 });
+  });
+});
+
+/**
+ * #635 Slice 0: live fallback must keep rollup/backfill careerCorrectSlots
+ * so avg correct still renders when freshness short-circuit fails.
+ */
+describe('withCareerCorrectSlots', () => {
+  const BASE = { totalPoints: 42, shows: 5, wins: 2 };
+
+  it('attaches a finite careerCorrectSlots from the user doc', () => {
+    expect(
+      withCareerCorrectSlots(BASE, { careerCorrectSlots: 18 })
+    ).toEqual({ ...BASE, careerCorrectSlots: 18 });
+  });
+
+  it('preserves zero careerCorrectSlots', () => {
+    expect(
+      withCareerCorrectSlots(BASE, { careerCorrectSlots: 0 })
+    ).toEqual({ ...BASE, careerCorrectSlots: 0 });
+  });
+
+  it('leaves stats unchanged when careerCorrectSlots is missing or invalid', () => {
+    expect(withCareerCorrectSlots(BASE, null)).toEqual(BASE);
+    expect(withCareerCorrectSlots(BASE, {})).toEqual(BASE);
+    expect(
+      withCareerCorrectSlots(BASE, { careerCorrectSlots: '18' })
+    ).toEqual(BASE);
+    expect(
+      withCareerCorrectSlots(BASE, { careerCorrectSlots: Number.NaN })
+    ).toEqual(BASE);
+  });
+
+  it('does not mutate the input stats object', () => {
+    const input = { ...BASE };
+    withCareerCorrectSlots(input, { careerCorrectSlots: 9 });
+    expect(input).toEqual(BASE);
+  });
+});
+
+/**
+ * #635 Slice 0: exclude today so an ungraded show-day calendar entry does
+ * not falsely stale rollups through yesterday.
+ */
+describe('deriveLatestFinalizedShow', () => {
+  it('returns the max calendar date strictly before today', () => {
+    expect(
+      deriveLatestFinalizedShow(
+        [
+          { date: '2026-07-15' },
+          { date: '2026-07-16' },
+          { date: '2026-07-17' },
+        ],
+        '2026-07-17'
+      )
+    ).toBe('2026-07-16');
+  });
+
+  it('skips today even when it is the only calendar date', () => {
+    expect(
+      deriveLatestFinalizedShow([{ date: '2026-07-17' }], '2026-07-17')
+    ).toBeNull();
+  });
+
+  it('skips future dates and empty/invalid entries', () => {
+    expect(
+      deriveLatestFinalizedShow(
+        [
+          { date: '2026-07-14' },
+          { date: '' },
+          null,
+          { date: '2026-07-20' },
+          { date: '2026-07-15' },
+        ],
+        '2026-07-17'
+      )
+    ).toBe('2026-07-15');
+  });
+
+  it('returns null for an empty calendar', () => {
+    expect(deriveLatestFinalizedShow([], '2026-07-17')).toBeNull();
+    expect(deriveLatestFinalizedShow(null, '2026-07-17')).toBeNull();
   });
 });
